W-14261340: MUnit can't start embedded container since modularization of mule-module-launcher
