+++ modules/launcher/src/main/java/module-info.java
@@ -27,8 +27,9 @@ module org.mule.runtime.launcher {
   requires org.mule.runtime.troubleshooting;
   requires org.mule.runtime.tooling.support;
 
-  exports org.mule.runtime.module.launcher to
-      org.mule.distribution.embedded.impl;
+  // Embedded uses this
+  exports org.mule.runtime.module.launcher;
+
   exports org.mule.runtime.module.launcher.coreextension to
       org.mule.runtime.core,
       com.mulesoft.mule.runtime.plugin;
