+++ modules/launcher/src/main/java/org/mule/runtime/module/launcher/DefaultMuleContainer.java
@@ -53,6 +53,7 @@ import org.mule.runtime.module.launcher.coreextension.ClasspathMuleCoreExtension
 import org.mule.runtime.module.launcher.coreextension.DefaultMuleCoreExtensionManagerServer;
 import org.mule.runtime.module.launcher.coreextension.MuleCoreExtensionManagerServer;
 import org.mule.runtime.module.launcher.coreextension.ReflectionMuleCoreExtensionDependencyResolver;
+import org.mule.runtime.module.launcher.splash.MuleContainerStartupSplashScreen;
 import org.mule.runtime.module.log4j.boot.api.MuleLog4jContextFactory;
 import org.mule.runtime.module.repository.api.RepositoryService;
 import org.mule.runtime.module.repository.internal.RepositoryServiceFactory;
